Removing the legacy FTP Backend
For this change, what was known as the alternate FTP client which uses FluentFTP library now officially replaces the FTP Backend which still used the deprecated FtpWebrequest.

To mantain compatibility with existing configurations, both ftp and aftp targets and configuration prefixes are accepted.
