+++ Duplicati/Library/Backend/FTP/FTPBackend.cs
@@ -19,444 +19,562 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER 
 // DEALINGS IN THE SOFTWARE.
 
-using Duplicati.Library.Common.IO;
-using Duplicati.Library.Interface;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+using System.IO;
 using System.Linq;
-using System.Text.RegularExpressions;
+using System.Net;
+using System.Net.Security;
+using System.Security.Authentication;
+using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using Duplicati.Library.Common.IO;
+using Duplicati.Library.Interface;
+using Duplicati.Library.Utility;
+using FluentFTP;
+using FluentFTP.Client.BaseClient;
+using FluentFTP.Exceptions;
+using CoreUtility = Duplicati.Library.Utility.Utility;
+using Uri = System.Uri;
 
 namespace Duplicati.Library.Backend
 {
-    // ReSharper disable once UnusedMember.Global
-    // This class is instantiated dynamically in the BackendLoader.
-    public class FTP : IBackend, IStreamingBackend
+    
+    /// <summary>
+    /// The unified FTP backend which uses the FluentFTP library.
+    ///
+    /// In previous versions, this was being exposed as AlternateFTPBackend, whist the FTP backend was
+    /// using the System.Net.FtpWebRequest class which is now deprecated.
+    ///
+    /// To provide a transparent upgrade path, the AlternateFTPBackend now inherits from this class,
+    /// but overides the default configuration values to match the old names(prefixed with a) and the backedn
+    /// name being "aftp" rather than ftp.
+    /// </summary>
+    public class FTP : IStreamingBackend
     {
-        private System.Net.NetworkCredential m_userInfo;
-        private readonly string m_url;
+        private NetworkCredential _userInfo;
+        private const string OPTION_ACCEPT_SPECIFIED_CERTIFICATE = "accept-specified-ssl-hash"; // Global option
+        private const string OPTION_ACCEPT_ANY_CERTIFICATE = "accept-any-ssl-certificate"; // Global option
 
-        private readonly bool m_useSSL = false;
-        private readonly bool m_passiveMode = false;
-        private readonly bool m_listVerify = true;
+        private const FtpDataConnectionType DEFAULT_DATA_CONNECTION_TYPE = FtpDataConnectionType.AutoPassive;
+        private const FtpEncryptionMode DEFAULT_ENCRYPTION_MODE = FtpEncryptionMode.None;
+        private static readonly SslProtocols DEFAULT_SSL_PROTOCOLS = SslProtocols.None; // NOTE: None means "use system default"
 
+        protected virtual string CONFIG_KEY_FTP_ENCRYPTION_MODE => "ftp-encryption-mode";
+        protected virtual string CONFIG_KEY_FTP_DATA_CONNECTION_TYPE => "ftp-data-connection-type";
+        protected virtual string CONFIG_KEY_FTP_SSL_PROTOCOLS => "ftp-ssl-protocols";
+        protected virtual string CONFIG_KEY_FTP_UPLOAD_DELAY => "ftp-upload-delay";
+        protected virtual string CONFIG_KEY_FTP_LOGTOCONSOLE => "ftp-log-to-console";
+        protected virtual string CONFIG_KEY_FTP_LOGPRIVATEINFOTOCONSOLE => "ftp-log-privateinfo-to-console";
+
+        private const string TEST_FILE_NAME = "duplicati-access-privileges-test.tmp";
+        private const string TEST_FILE_CONTENT = "This file is used by Duplicati to test access permissions and can be safely deleted.";
+
+        // ReSharper disable InconsistentNaming
+        private static readonly string DEFAULT_DATA_CONNECTION_TYPE_STRING = DEFAULT_DATA_CONNECTION_TYPE.ToString();
+        private static readonly string DEFAULT_ENCRYPTION_MODE_STRING = DEFAULT_ENCRYPTION_MODE.ToString();
+        private static readonly string DEFAULT_SSL_PROTOCOLS_STRING = DEFAULT_SSL_PROTOCOLS.ToString();
+        private static readonly string DEFAULT_UPLOAD_DELAY_STRING = "0s";
+        // ReSharper restore InconsistentNaming
+
+        private readonly string _url;
+        private readonly bool _listVerify = true;
+        private readonly FtpConfig _ftpConfig;
+        private readonly TimeSpan _uploadWaitTime;
+
+        private readonly bool _logToConsole;
+        private readonly bool _logPrivateInfoToConsole;
+        private readonly bool _accepAllCertificates;
+        private readonly string[] _validHashes;
+
+        /// <summary>
+        /// The localized name to display for this backend
+        /// </summary>
+        public virtual string DisplayName => Strings.DisplayName;
+
+        /// <summary>
+        /// The protocol key, e.g. ftp, http or ssh
+        /// </summary>
+        public virtual string ProtocolKey => "ftp";
+
+        private AsyncFtpClient Client
+        { get; set; }
+
+        public IList<ICommandLineArgument> SupportedCommands =>
+            new List<ICommandLineArgument>([
+                new CommandLineArgument("auth-password", CommandLineArgument.ArgumentType.Password, Strings.DescriptionAuthPasswordShort, Strings.DescriptionAuthPasswordLong),
+                new CommandLineArgument("auth-username", CommandLineArgument.ArgumentType.String, Strings.DescriptionAuthUsernameShort, Strings.DescriptionAuthUsernameLong),
+                new CommandLineArgument("disable-upload-verify", CommandLineArgument.ArgumentType.Boolean, Strings.DescriptionDisableUploadVerifyShort, Strings.DescriptionDisableUploadVerifyLong),
+                new CommandLineArgument(CONFIG_KEY_FTP_DATA_CONNECTION_TYPE, CommandLineArgument.ArgumentType.Enumeration, Strings.DescriptionFtpDataConnectionTypeShort, Strings.DescriptionFtpDataConnectionTypeLong, DEFAULT_DATA_CONNECTION_TYPE_STRING, null, Enum.GetNames(typeof(FtpDataConnectionType))),
+                new CommandLineArgument(CONFIG_KEY_FTP_ENCRYPTION_MODE, CommandLineArgument.ArgumentType.Enumeration, Strings.DescriptionFtpEncryptionModeShort, Strings.DescriptionFtpEncryptionModeLong, DEFAULT_ENCRYPTION_MODE_STRING, null, Enum.GetNames(typeof(FtpEncryptionMode))),
+                new CommandLineArgument(CONFIG_KEY_FTP_SSL_PROTOCOLS, CommandLineArgument.ArgumentType.Flags, Strings.DescriptionSslProtocolsShort, Strings.DescriptionSslProtocolsLong, DEFAULT_SSL_PROTOCOLS_STRING, null, Enum.GetNames(typeof(SslProtocols))),
+                new CommandLineArgument(CONFIG_KEY_FTP_UPLOAD_DELAY, CommandLineArgument.ArgumentType.Timespan, Strings.DescriptionUploadDelayShort, Strings.DescriptionUploadDelayLong, DEFAULT_UPLOAD_DELAY_STRING),
+                new CommandLineArgument(CONFIG_KEY_FTP_LOGTOCONSOLE, CommandLineArgument.ArgumentType.Boolean, Strings.DescriptionLogToConsoleShort, Strings.DescriptionLogToConsoleLong),
+                new CommandLineArgument(CONFIG_KEY_FTP_LOGPRIVATEINFOTOCONSOLE, CommandLineArgument.ArgumentType.Boolean, Strings.DescriptionLogPrivateInfoToConsoleShort, Strings.DescriptionLogPrivateInfoToConsoleLong, "false"),
+            ]);
+
+        /// <summary>
+        /// Initialize a new instance.
+        /// </summary>
         public FTP()
         {
+
         }
 
+        /// <summary>
+        /// Initialize a new instance/
+        /// </summary>
+        /// <param name="url">Configured url.</param>
+        /// <param name="options">Configured options. cannot be null.</param>
+        [SuppressMessage("ReSharper", "VirtualMemberCallInConstructor")] // The behavior of accessing the virtual properties is as expected
         public FTP(string url, Dictionary<string, string> options)
         {
-            //This can be made better by keeping a single ftp stream open,
-            //unfortunately the .Net model does not allow this as the request is 
-            //bound to a single url (path+file).
-            //
-            //To fix this, a thirdparty FTP library is required,
-            //this would also allow a fix for the FTP servers
-            //that only support SSL during authentication, not during transfers
-            //
-            //If you have experience with a stable open source .Net FTP library,
-            //please let the Duplicati authors know
+            _accepAllCertificates = CoreUtility.ParseBoolOption(options, OPTION_ACCEPT_ANY_CERTIFICATE);
+
+            string certHash;
+            options.TryGetValue(OPTION_ACCEPT_SPECIFIED_CERTIFICATE, out certHash);
+
+            _validHashes = certHash?.Split([",", ";"], StringSplitOptions.RemoveEmptyEntries);
 
             var u = new Utility.Uri(url);
             u.RequireHost();
-            string username = null;
-            string password = null;
 
             if (!string.IsNullOrEmpty(u.Username))
             {
-                username = u.Username;
-            }
-            else if (options.ContainsKey("auth-username"))
-            {
-                username = options["auth-username"];
-            }
-
-            if (!string.IsNullOrEmpty(u.Username) && !string.IsNullOrEmpty(u.Password))
-            {
-                password = u.Password;
-            }
-            else if (options.ContainsKey("auth-password"))
-            {
-                password = options["auth-password"];
-            }
-
-            m_userInfo = new System.Net.NetworkCredential
-            {
-                UserName = username,
-                Password = password
-            };
-
-            //Bugfix, see http://connect.microsoft.com/VisualStudio/feedback/details/695227/networkcredential-default-constructor-leaves-domain-null-leading-to-null-object-reference-exceptions-in-framework-code
-            if (m_userInfo != null)
-                m_userInfo.Domain = "";
-
-            m_url = u.SetScheme("ftp").SetQuery(null).SetCredentials(null, null).ToString();
-            m_url = Util.AppendDirSeparator(m_url, "/");
-
-            m_useSSL = Utility.Utility.ParseBoolOption(options, "use-ssl");
-
-            m_listVerify = !Utility.Utility.ParseBoolOption(options, "disable-upload-verify");
-
-            if (Utility.Utility.ParseBoolOption(options, "ftp-passive"))
-            {
-                m_passiveMode = true;
-            }
-            else m_passiveMode = !Utility.Utility.ParseBoolOption(options, "ftp-regular");
-        }
-
-        #region Regular expression to parse list lines
-        //Regexps found here: http://www.dotnetfunda.com/articles/article125.aspx
-        //Modified to allow hyphens in username and groupname
-        internal readonly static Regex[] PARSEFORMATS =
-        {
-            new Regex(@"(?<dir>[\-d])(?<permission>([\-r][\-w][\-xs]){3})\s+\d+\s+(?<groupname>\S+)\s+(?<username>\S+)\s+(?<size>\d+)\s+(?<timestamp>\w+\s+\d+\s+\d{4})\s+(?<name>.+)"),
-            new Regex(@"(?<dir>[\-d])(?<permission>([\-r][\-w][\-xs]){3})\s+(?<groupname>\d+)\s+(?<username>\d+)\s+(?<size>\d+)\s+(?<timestamp>\w+\s+\d+\s+\d{4})\s+(?<name>.+)"),
-            new Regex(@"(?<dir>[\-d])(?<permission>([\-r][\-w][\-xs]){3})\s+(?<groupname>\d+)\s+(?<username>\d+)\s+(?<size>\d+)\s+(?<timestamp>\w+\s+\d+\s+\d{1,2}:\d{2})\s+(?<name>.+)"),
-            new Regex(@"(?<dir>[\-d])(?<permission>([\-r][\-w][\-xs]){3})\s+\d+\s+(?<groupname>\S+)\s+(?<username>\S+)\s+(?<size>\d+)\s+(?<timestamp>\w+\s+\d+\s+\d{1,2}:\d{2})\s+(?<name>.+)"),
-            new Regex(@"(?<dir>[\-d])(?<permission>([\-r][\-w][\-xs]){3})(\s+)(?<size>(\d+))(\s+)(?<ctbit>(\w+\s\w+))(\s+)(?<size2>(\d+))\s+(?<timestamp>\w+\s+\d+\s+\d{2}:\d{2})\s+(?<name>.+)"),
-            new Regex(@"(?<timestamp>\d{2}\-\d{2}\-\d{2}\s+\d{2}:\d{2}[Aa|Pp][mM])\s+(?<dir>\<\w+\>){0,1}(?<size>\d+){0,1}\s+(?<name>.+)"),
-            new Regex(@"([<timestamp>]*\d{2}\-\d{2}\-\d{2}\s+\d{2}:\d{2}[Aa|Pp][mM])\s+([<dir>]*\<\w+\>){0,1}([<size>]*\d+){0,1}\s+([<name>]*.+)")
-        };
-        #endregion
-
-        private static Match MatchLine(string line)
-        {
-            Match m = null;
-            foreach (Regex s in PARSEFORMATS)
-                if ((m = s.Match(line)).Success)
-                    return m;
-
-            return null;
-        }
-
-        public static FileEntry ParseLine(string line)
-        {
-            Match m = MatchLine(line);
-            if (m == null)
-                return null;
-
-            FileEntry f = new FileEntry(m.Groups["name"].Value);
-
-            string time = m.Groups["timestamp"].Value;
-            string dir = m.Groups["dir"].Value;
-
-            if (dir != "" && dir != "-")
-            {
-                f.IsFolder = true;
+                _userInfo = new NetworkCredential
+                {
+                    UserName = u.Username
+                };
+                if (!string.IsNullOrEmpty(u.Password))
+                    _userInfo.Password = u.Password;
+                else if (options.ContainsKey("auth-password"))
+                    _userInfo.Password = options["auth-password"];
             }
             else
             {
-                f.Size = long.Parse(m.Groups["size"].Value);
+                if (options.ContainsKey("auth-username"))
+                {
+                    _userInfo = new NetworkCredential();
+                    _userInfo.UserName = options["auth-username"];
+                    if (options.ContainsKey("auth-password"))
+                        _userInfo.Password = options["auth-password"];
+                }
             }
 
-            if (DateTime.TryParse(time, out DateTime t))
+            //Bugfix, see http://connect.microsoft.com/VisualStudio/feedback/details/695227/networkcredential-default-constructor-leaves-domain-null-leading-to-null-object-reference-exceptions-in-framework-code
+            if (_userInfo != null)
+                _userInfo.Domain = "";
+
+            _url = u.SetScheme("ftp").SetQuery(null).SetCredentials(null, null).ToString();
+            _url = Util.AppendDirSeparator(_url, "/");
+            _listVerify = !CoreUtility.ParseBoolOption(options, "disable-upload-verify");
+            if (options.TryGetValue(CONFIG_KEY_FTP_UPLOAD_DELAY, out var uploadWaitTimeString) && !string.IsNullOrWhiteSpace(uploadWaitTimeString))
+                _uploadWaitTime = Timeparser.ParseTimeSpan(uploadWaitTimeString);
+
+            // Process the aftp-data-connection-type option
+            string dataConnectionTypeString;
+            FtpDataConnectionType dataConnectionType;
+
+            if (!options.TryGetValue(CONFIG_KEY_FTP_DATA_CONNECTION_TYPE, out dataConnectionTypeString) || string.IsNullOrWhiteSpace(dataConnectionTypeString))
+                dataConnectionTypeString = null;
+
+            if (dataConnectionTypeString == null || !Enum.TryParse(dataConnectionTypeString, true, out dataConnectionType))
+                dataConnectionType = DEFAULT_DATA_CONNECTION_TYPE;
+
+            // Process the aftp-encryption-mode option
+            FtpEncryptionMode encryptionMode;
+            if (!options.TryGetValue(CONFIG_KEY_FTP_ENCRYPTION_MODE, out var encryptionModeString) || string.IsNullOrWhiteSpace(encryptionModeString))
+                encryptionModeString = null;
+
+            if (encryptionModeString == null || !Enum.TryParse(encryptionModeString, true, out encryptionMode))
+                encryptionMode = DEFAULT_ENCRYPTION_MODE;
+
+            // Process the aftp-ssl-protocols option
+            SslProtocols sslProtocols;
+            if (!options.TryGetValue(CONFIG_KEY_FTP_SSL_PROTOCOLS, out var sslProtocolsString) || string.IsNullOrWhiteSpace(sslProtocolsString))
+                sslProtocolsString = null;
+
+            if (sslProtocolsString == null || !Enum.TryParse(sslProtocolsString, true, out sslProtocols)) sslProtocols = DEFAULT_SSL_PROTOCOLS;
+
+            _logToConsole = CoreUtility.ParseBoolOption(options, CONFIG_KEY_FTP_LOGTOCONSOLE);
+            _logPrivateInfoToConsole = CoreUtility.ParseBoolOption(options, CONFIG_KEY_FTP_LOGPRIVATEINFOTOCONSOLE);
+            
+            _ftpConfig = new FtpConfig
             {
-                f.LastAccess = f.LastModification = t;
-            }
+                DataConnectionType = dataConnectionType,
+                EncryptionMode = encryptionMode,
+                SslProtocols = sslProtocols,
+                LogToConsole = _logToConsole,
+            };
 
-            return f;
-        }
-
-        #region IBackendInterface Members
-
-        public string DisplayName
-        {
-            get { return Strings.FTPBackend.DisplayName; }
-        }
-
-        public string ProtocolKey
-        {
-            get { return "ftp"; }
-        }
-
-        private T HandleListExceptions<T>(Func<T> func, System.Net.FtpWebRequest req)
-        {
-            T ret = default(T);
-            Action action = () => ret = func();
-            HandleListExceptions(action, req);
-            return ret;
-        }
-
-        private void HandleListExceptions(Action action, System.Net.FtpWebRequest req)
-        {
-            try
-            {
-                action();
-            }
-            catch (System.Net.WebException wex)
-            {
-                if (wex.Response as System.Net.FtpWebResponse != null && (wex.Response as System.Net.FtpWebResponse).StatusCode == System.Net.FtpStatusCode.ActionNotTakenFileUnavailable)
-                    throw new Interface.FolderMissingException(Strings.FTPBackend.MissingFolderError(req.RequestUri.PathAndQuery, wex.Message), wex);
-                else
-                    throw;
-            }
+            if (_logPrivateInfoToConsole) _ftpConfig.LogHost = _ftpConfig.LogPassword = _ftpConfig.LogUserName = true;
         }
 
         public IEnumerable<IFileEntry> List()
         {
-            return List("");
+            return List(string.Empty);
         }
 
         public IEnumerable<IFileEntry> List(string filename)
         {
-            var req = CreateRequest(filename);
-            req.Method = System.Net.WebRequestMethods.Ftp.ListDirectoryDetails;
-            req.UseBinary = false;
-
-            System.Net.WebResponse resp = null;
-            System.IO.Stream rs = null;
-            System.IO.StreamReader sr = null;
-
-            try
-            {
-                HandleListExceptions(
-                    () =>
-                        {
-                            var areq = new Utility.AsyncHttpRequest(req);
-                            resp = areq.GetResponse();
-                            rs = areq.GetResponseStream();
-                            sr = new System.IO.StreamReader(new StreamReadHelper(rs));
-                        },
-                    req);
-
-                string line;
-                while ((line = HandleListExceptions(sr.ReadLine, req)) != null)
-                {
-                    FileEntry f = ParseLine(line);
-                    if (f != null)
-                        yield return f;
-                }
-            }
-            finally
-            {
-                try
-                {
-                    if (sr != null)
-                    {
-                        sr.Dispose();
-                    }
-                }
-                finally
-                {
-                    try
-                    {
-                        if (rs != null)
-                        {
-                            rs.Dispose();
-                        }
-                    }
-                    finally
-                    {
-                        if (resp != null)
-                        {
-                            resp.Dispose();
-                        }
-                    }
-                }
-            }
+            return List(filename, false);
         }
 
-        public async Task PutAsync(string remotename, System.IO.Stream input, CancellationToken cancelToken)
+        private IEnumerable<IFileEntry> List(string filename, bool stripFile)
         {
-            System.Net.FtpWebRequest req = null;
-            try
+            var list = new List<IFileEntry>();
+            string remotePath = filename;
+
+            var ftpClient = CreateClient();
+
+            // Get the remote path
+            var url = new Uri(_url);
+            remotePath = "/" + GetUnescapedAbsolutePath(url);
+
+            if (!string.IsNullOrEmpty(filename))
             {
-                req = CreateRequest(remotename);
-                req.Method = System.Net.WebRequestMethods.Ftp.UploadFile;
-                req.UseBinary = true;
-
-                long streamLen = -1;
-                try { streamLen = input.Length; }
-                catch { }
-
-                Utility.AsyncHttpRequest areq = new Utility.AsyncHttpRequest(req);
-                using (System.IO.Stream rs = areq.GetRequestStream(streamLen))
-                    await Utility.Utility.CopyStreamAsync(input, rs, true, cancelToken).ConfigureAwait(false);
-
-                if (m_listVerify)
+                if (!stripFile)
                 {
-                    IEnumerable<IFileEntry> files = List(remotename);
-                    foreach (var fe in files)
-                        if (fe.Name.Equals(remotename) || fe.Name.EndsWith("/" + remotename, StringComparison.Ordinal) || fe.Name.EndsWith("\\" + remotename, StringComparison.Ordinal))
-                        {
-                            if (fe.Size < 0 || streamLen < 0 || fe.Size == streamLen)
-                                return;
+                    // Append the filename
+                    remotePath += filename;
+                }
+                else if (filename.Contains("/"))
+                {
+                    remotePath += filename.Substring(0, filename.LastIndexOf("/", StringComparison.Ordinal));
+                }
+                // else: stripping the filename in this case ignoring it
+            }
 
-                            throw new Exception(Strings.FTPBackend.ListVerifySizeFailure(remotename, fe.Size, streamLen));
+            foreach (FtpListItem item in ftpClient.GetListing(remotePath, FtpListOption.Modify | FtpListOption.Size).Await())
+            {
+                switch (item.Type)
+                {
+                    case FtpObjectType.Directory:
+                        {
+                            if (item.Name == "." || item.Name == "..")
+                            {
+                                continue;
+                            }
+
+                            list.Add(new FileEntry(item.Name, -1, new DateTime(), item.Modified)
+                            {
+                                IsFolder = true,
+                            });
+
+                            break;
+                        }
+                    case FtpObjectType.File:
+                        {
+                            list.Add(new FileEntry(item.Name, item.Size, new DateTime(), item.Modified));
+
+                            break;
+                        }
+                    case FtpObjectType.Link:
+                        {
+                            if (item.Name == "." || item.Name == "..")
+                            {
+                                continue;
+                            }
+
+                            if (item.LinkObject != null)
+                            {
+                                switch (item.LinkObject.Type)
+                                {
+                                    case FtpObjectType.Directory:
+                                        {
+                                            if (item.Name == "." || item.Name == "..")
+                                            {
+                                                continue;
+                                            }
+
+                                            list.Add(new FileEntry(item.Name, -1, new DateTime(), item.Modified)
+                                            {
+                                                IsFolder = true,
+                                            });
+
+                                            break;
+                                        }
+                                    case FtpObjectType.File:
+                                        {
+                                            list.Add(new FileEntry(item.Name, item.Size, new DateTime(), item.Modified));
+
+                                            break;
+                                        }
+                                }
+                            }
+                            break;
                         }
 
-                    throw new Exception(Strings.FTPBackend.ListVerifyFailure(remotename, files.Select(n => n.Name)));
                 }
-
             }
-            catch (System.Net.WebException wex)
+
+            return list;
+        }
+
+        public async Task PutAsync(string remotename, Stream input, CancellationToken cancelToken)
+        {
+            string remotePath = remotename;
+            long streamLen;
+
+            var ftpClient = CreateClient();
+
+            try
             {
-                if (req != null && wex.Response as System.Net.FtpWebResponse != null && (wex.Response as System.Net.FtpWebResponse).StatusCode == System.Net.FtpStatusCode.ActionNotTakenFileUnavailable)
-                    throw new Interface.FolderMissingException(Strings.FTPBackend.MissingFolderError(req.RequestUri.PathAndQuery, wex.Message), wex);
-                else
-                    throw;
+                streamLen = input.Length;
+            }
+            catch (NotSupportedException) { streamLen = -1; }
+
+            // Get the remote path
+            remotePath = "";
+
+            if (!string.IsNullOrEmpty(remotename))
+            {
+                // Append the filename
+                remotePath += remotename;
+            }
+
+            var status = await ftpClient.UploadStream(input, remotePath, createRemoteDir: false, token: cancelToken, progress: null).ConfigureAwait(false);
+            if (status != FtpStatus.Success)
+            {
+                throw new UserInformationException(string.Format(Strings.ErrorWriteFile, remotename), "AftpPutFailure");
+            }
+
+            // Wait for the upload, if required
+            if (_uploadWaitTime.Ticks > 0)
+            {
+                Thread.Sleep(_uploadWaitTime);
+            }
+
+            if (_listVerify)
+            {
+                // check remote file size; matching file size indicates completion
+                var remoteSize = await ftpClient.GetFileSize(remotePath, -1, cancelToken);
+                if (streamLen != remoteSize)
+                {
+                    throw new UserInformationException(Strings.ListVerifySizeFailure(remotename, remoteSize, streamLen), "AftpListVerifySizeFailure");
+                }
             }
         }
 
         public async Task PutAsync(string remotename, string localname, CancellationToken cancelToken)
         {
-            using (System.IO.FileStream fs = System.IO.File.Open(localname, System.IO.FileMode.Open, System.IO.FileAccess.Read, System.IO.FileShare.Read))
-                await PutAsync(remotename, fs, cancelToken).ConfigureAwait(false);
+            await using FileStream fs = File.Open(localname, FileMode.Open, FileAccess.Read, FileShare.Read);
+            await PutAsync(remotename, fs, cancelToken);
         }
 
-        public async Task GetAsync(string remotename, System.IO.Stream output, CancellationToken cancelToken)
+        public async Task GetAsync(string remotename, Stream output, CancellationToken cancelToken)
         {
-            var req = CreateRequest(remotename);
-            req.Method = System.Net.WebRequestMethods.Ftp.DownloadFile;
-            req.UseBinary = true;
+            var ftpClient = CreateClient();
 
-            var areq = new Utility.AsyncHttpRequest(req);
-            using (var resp = areq.GetResponse())
-            using (var rs = areq.GetResponseStream())
-                await Utility.Utility.CopyStreamAsync(rs, output, false, cancelToken).ConfigureAwait(false);
+            // Get the remote path
+            var remotePath = "";
+
+            if (!string.IsNullOrEmpty(remotename))
+            {
+                // Append the filename
+                remotePath += remotename;
+            }
+
+            await using var inputStream = await ftpClient.OpenRead(remotePath, token: cancelToken);
+            try
+            {
+                await CoreUtility.CopyStreamAsync(inputStream, output, false, cancelToken).ConfigureAwait(false);
+            }
+            finally
+            {
+                inputStream.Close();
+            }
         }
 
         public async Task GetAsync(string remotename, string localname, CancellationToken cancelToken)
         {
-            using (System.IO.FileStream fs = System.IO.File.Open(localname, System.IO.FileMode.Create, System.IO.FileAccess.Write, System.IO.FileShare.None))
-                await GetAsync(remotename, fs, cancelToken).ConfigureAwait(false);
+            await using FileStream fs = File.Open(localname, FileMode.Create, FileAccess.Write, FileShare.None);
+            await GetAsync(remotename, fs, cancelToken).ConfigureAwait(false);
         }
 
-        public Task DeleteAsync(string remotename, CancellationToken cancelToken)
+        public async Task DeleteAsync(string remotename, CancellationToken cancelToken)
         {
-            System.Net.FtpWebRequest req = CreateRequest(remotename);
-            req.Method = System.Net.WebRequestMethods.Ftp.DeleteFile;
-            Utility.AsyncHttpRequest areq = new Utility.AsyncHttpRequest(req);
-            using (areq.GetResponse())
-            { }
+            var ftpClient = CreateClient();
 
-            return Task.CompletedTask;
-        }
+            // Get the remote path
+            var remotePath = "";
 
-        public IList<ICommandLineArgument> SupportedCommands
-        {
-            get
+            if (!string.IsNullOrEmpty(remotename))
             {
-                return new List<ICommandLineArgument>(new ICommandLineArgument[] {
-                    new CommandLineArgument("ftp-passive", CommandLineArgument.ArgumentType.Boolean, Strings.FTPBackend.DescriptionFTPPassiveShort, Strings.FTPBackend.DescriptionFTPPassiveLong, "false"),
-                    new CommandLineArgument("ftp-regular", CommandLineArgument.ArgumentType.Boolean, Strings.FTPBackend.DescriptionFTPActiveShort, Strings.FTPBackend.DescriptionFTPActiveLong, "true"),
-                    new CommandLineArgument("auth-password", CommandLineArgument.ArgumentType.Password, Strings.FTPBackend.DescriptionAuthPasswordShort, Strings.FTPBackend.DescriptionAuthPasswordLong),
-                    new CommandLineArgument("auth-username", CommandLineArgument.ArgumentType.String, Strings.FTPBackend.DescriptionAuthUsernameShort, Strings.FTPBackend.DescriptionAuthUsernameLong),
-                    new CommandLineArgument("use-ssl", CommandLineArgument.ArgumentType.Boolean, Strings.FTPBackend.DescriptionUseSSLShort, Strings.FTPBackend.DescriptionUseSSLLong),
-                    new CommandLineArgument("disable-upload-verify", CommandLineArgument.ArgumentType.Boolean, Strings.FTPBackend.DescriptionDisableUploadVerifyShort, Strings.FTPBackend.DescriptionDisableUploadVerifyLong),
-                });
-            }
-        }
-
-        public string Description
-        {
-            get
-            {
-                return Strings.FTPBackend.Description;
-            }
-        }
-
-        public Task<string[]> GetDNSNamesAsync(CancellationToken cancelToken) => Task.FromResult(new string[] { new Uri(m_url).Host });
-
-        public Task TestAsync(CancellationToken cancelToken)
-        {
-            this.TestList();
-            return Task.CompletedTask;
-        }
-
-        public Task CreateFolderAsync(CancellationToken cancelToken)
-        {
-            System.Net.FtpWebRequest req = CreateRequest("", true);
-            req.Method = System.Net.WebRequestMethods.Ftp.MakeDirectory;
-            req.KeepAlive = false;
-            Utility.AsyncHttpRequest areq = new Utility.AsyncHttpRequest(req);
-            using (areq.GetResponse())
-            { }
-
-            return Task.CompletedTask;
-        }
-
-        #endregion
-
-        #region IDisposable Members
-
-        public void Dispose()
-        {
-            m_userInfo = null;
-        }
-
-        #endregion
-
-        private System.Net.FtpWebRequest CreateRequest(string remotename)
-        {
-            return CreateRequest(remotename, false);
-        }
-
-        private System.Net.FtpWebRequest CreateRequest(string remotename, bool createFolder)
-        {
-            string url = m_url;
-            if (createFolder && url.EndsWith("/", StringComparison.Ordinal))
-                url = url.Substring(0, url.Length - 1);
-
-            System.Net.FtpWebRequest req = (System.Net.FtpWebRequest)System.Net.WebRequest.Create(url + remotename);
-
-            if (m_userInfo != null)
-            {
-                req.Credentials = m_userInfo;
+                // Append the filename
+                remotePath += remotename;
             }
 
-            req.KeepAlive = false;
-            req.UsePassive = m_passiveMode;
+            await ftpClient.DeleteFile(remotePath, cancelToken);
 
-            if (m_useSSL)
-                req.EnableSsl = m_useSSL;
-
-            return req;
         }
 
         /// <summary>
-        /// Private helper class to fix a bug with the StreamReader
+        /// A localized description of the backend, for display in the usage information
         /// </summary>
-        private class StreamReadHelper : Utility.OverrideableStream
-        {
-            /// <summary>
-            /// Once the stream has returned 0 as the read count it is disposed
-            /// in the FtpRequest, and subsequent read requests will throw an ObjectDisposedException
-            /// </summary>
-            private bool m_empty = false;
+        public virtual string Description => Strings.Description;
 
-            /// <summary>
-            /// Basic initialization, just pass the stream to the super class
-            /// </summary>
-            /// <param name="stream"></param>
-            public StreamReadHelper(System.IO.Stream stream)
-                : base(stream)
+        public Task<string[]> GetDNSNamesAsync(CancellationToken cancelToken) => Task.FromResult(new[] { new Uri(_url).Host });
+
+        /// <summary>
+        /// Test FTP access permissions.
+        /// </summary>
+        public async Task TestAsync(CancellationToken cancellationToken)
+        {
+            var list = List();
+
+            // Delete test file if exists
+            if (list.Any(entry => entry.Name == TEST_FILE_NAME))
             {
+                try
+                {
+                    await DeleteAsync(TEST_FILE_NAME, cancellationToken).ConfigureAwait(false);
+                }
+                catch (Exception e)
+                {
+                    if (e.InnerException != null) { e = e.InnerException; }
+                    throw new Exception(string.Format(Strings.ErrorDeleteFile, e.Message), e);
+                }
             }
 
-            /// <summary>
-            /// Override the read function to make sure that we only return less than the requested amount of data if the stream is exhausted
-            /// </summary>
-            /// <param name="buffer">The buffer to place data in</param>
-            /// <param name="offset">The offset into the buffer to start at</param>
-            /// <param name="count">The number of bytes to read</param>
-            /// <returns>The number of bytes read</returns>
-            public override int Read(byte[] buffer, int offset, int count)
+            // Test write permissions
+            using (var testStream = new MemoryStream(Encoding.UTF8.GetBytes(TEST_FILE_CONTENT)))
             {
-                int readCount = 0;
-                int a;
-
-                while (!m_empty && count > 0)
+                try
                 {
-                    a = base.Read(buffer, offset, count);
-                    readCount += a;
-                    count -= a;
-                    offset += a;
-                    m_empty = a == 0;
+                    await PutAsync(TEST_FILE_NAME, testStream, cancellationToken).ConfigureAwait(false);
                 }
+                catch (Exception e)
+                {
+                    if (e.InnerException != null) { e = e.InnerException; }
+                    throw new Exception(string.Format(Strings.ErrorWriteFile, e.Message), e);
+                }
+            }
 
-                return readCount;
+            // Test read permissions
+            using (var testStream = new MemoryStream())
+            {
+                try
+                {
+                    await GetAsync(TEST_FILE_NAME, testStream, cancellationToken).ConfigureAwait(false);
+                    var readValue = Encoding.UTF8.GetString(testStream.ToArray());
+                    if (readValue != TEST_FILE_CONTENT)
+                        throw new Exception("Test file corrupted.");
+                }
+                catch (Exception e)
+                {
+                    if (e.InnerException != null) { e = e.InnerException; }
+                    throw new Exception(string.Format(Strings.ErrorReadFile, e.Message), e);
+                }
+            }
+
+            // Cleanup
+            try
+            {
+                await DeleteAsync(TEST_FILE_NAME, cancellationToken).ConfigureAwait(false);
+            }
+            catch (Exception e)
+            {
+                if (e.InnerException != null) { e = e.InnerException; }
+                throw new Exception(string.Format(Strings.ErrorDeleteFile, e.Message), e);
+            }
+        }
+
+        public Task CreateFolderAsync(CancellationToken cancellationToken)
+        {
+            var client = CreateClient(false);
+
+            var url = new Uri(_url);
+
+            // Get the remote path
+            var remotePath = GetUnescapedAbsolutePath(url);
+
+            // Try to create the directory 
+            return client.CreateDirectory(remotePath, true, cancellationToken);
+        }
+
+        public void Dispose()
+        {
+            if (Client != null)
+                Client.Dispose();
+
+            Client = null;
+            _userInfo = null;
+        }
+
+        private AsyncFtpClient CreateClient(bool setWorkingDirectory = true)
+        {
+            var uri = new Uri(_url);
+
+            if (Client == null) // Create connection if it doesn't exist yet
+            {
+                var ftpClient = new AsyncFtpClient
+                {
+                    Host = uri.Host,
+                    Port = uri.Port == -1 ? 21 : uri.Port,
+                    Credentials = _userInfo,
+                    Config = _ftpConfig,
+                };
+
+                ftpClient.ValidateCertificate += HandleValidateCertificate;
+
+                Client = ftpClient;
+
+            } // else reuse existing connection
+
+            if (setWorkingDirectory)
+            {
+                // Change working directory to the remote path
+                // Do this every time to prevent issues when FtpClient silently reconnects after failure.
+                var remotePath = GetUnescapedAbsolutePath(uri);
+                try
+                {
+                    Client.SetWorkingDirectory(remotePath).Await();
+                }
+                catch (FtpCommandException ex)
+                {
+                    if (ex.CompletionCode == "550")
+                    {
+                        throw new FolderMissingException(Strings.MissingFolderError(remotePath, ex.Message), ex);
+                    }
+
+                    throw;
+                }
+            }
+
+            return Client;
+        }
+
+        private string GetUnescapedAbsolutePath(Uri uri)
+        {
+            string absolutePath = Uri.UnescapeDataString(uri.AbsolutePath);
+            return absolutePath.EndsWith("/", StringComparison.Ordinal) ? absolutePath.Substring(0, absolutePath.Length - 1) : absolutePath;
+        }
+
+        private void HandleValidateCertificate(BaseFtpClient control, FtpSslValidationEventArgs e)
+        {
+            if (e.PolicyErrors == SslPolicyErrors.None || _accepAllCertificates)
+            {
+                e.Accept = true;
+                return;
+            }
+
+            try
+            {
+                var certHash = (_validHashes != null && _validHashes.Length > 0) ? CoreUtility.ByteArrayAsHexString(e.Certificate.GetCertHash()) : null;
+                if (certHash != null)
+                {
+                    if (_validHashes.Any(hash => !string.IsNullOrEmpty(hash) && certHash.Equals(hash, StringComparison.OrdinalIgnoreCase)))
+                    {
+                        e.Accept = true;
+                    }
+                }
+            }
+            catch
+            {
+                e.Accept = false;
             }
         }
     }
